move process utt to collator
