+++ examples/tiny/s0/conf/deepspeech2.yaml
@@ -6,7 +6,7 @@ data:
   mean_std_filepath: data/mean_std.json
   vocab_filepath: data/vocab.txt 
   augmentation_config: conf/augmentation.json
-  batch_size: 4
+  batch_size: 2
   min_input_len: 0.0
   max_input_len: 27.0
   min_output_len: 0.0
@@ -37,7 +37,7 @@ model:
   share_rnn_weights: True 
 
 training:
-  n_epoch: 20
+  n_epoch: 10
   lr: 1e-5 
   lr_decay: 1.0 
   weight_decay: 1e-06
